Created utils package for shared code
