+++ functions/backfill/post/main.go
@@ -10,16 +10,16 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 
-	dynamo_types "github.com/KamWithK/exSTATic-backend"
+	"github.com/KamWithK/exSTATic-backend/utils"
 )
 
 var sess *session.Session
 var svc *dynamodb.DynamoDB
 
 type BackfillArgs struct {
-	Username     string                        `json:"username"`
-	MediaEntries []dynamo_types.UserMediaEntry `json:"media_entries"`
-	MediaStats   []dynamo_types.UserMediaStat  `json:"media_stats"`
+	Username     string                 `json:"username"`
+	MediaEntries []utils.UserMediaEntry `json:"media_entries"`
+	MediaStats   []utils.UserMediaStat  `json:"media_stats"`
 }
 
 func init() {
@@ -30,12 +30,12 @@ func init() {
 }
 
 func putRequest(pk string, sk string, userMedia interface{}) *dynamodb.WriteRequest {
-	tableKey, keyErr := dynamo_types.GetCompositeKey(pk, sk)
+	tableKey, keyErr := utils.GetCompositeKey(pk, sk)
 	if keyErr != nil {
 		return nil
 	}
 
-	writeRequest, writeErr := dynamo_types.PutItemRequest(svc, tableKey, userMedia)
+	writeRequest, writeErr := utils.PutItemRequest(svc, tableKey, userMedia)
 
 	if writeErr != nil {
 		return nil
@@ -44,7 +44,7 @@ func putRequest(pk string, sk string, userMedia interface{}) *dynamodb.WriteRequ
 	return writeRequest
 }
 
-func HandleRequest(ctx context.Context, history BackfillArgs) (*dynamo_types.BatchwriteArgs, error) {
+func HandleRequest(ctx context.Context, history BackfillArgs) (*utils.BatchwriteArgs, error) {
 	username := history.Username
 
 	if len(username) == 0 {
@@ -67,7 +67,7 @@ func HandleRequest(ctx context.Context, history BackfillArgs) (*dynamo_types.Bat
 	}
 
 	for _, userMedia := range history.MediaStats {
-		writeRequest := putRequest(userMedia.Key.MediaType+"#"+username, dynamo_types.ZeroPadInt64(*userMedia.Date)+"#"+userMedia.Key.MediaIdentifier, &userMedia)
+		writeRequest := putRequest(userMedia.Key.MediaType+"#"+username, utils.ZeroPadInt64(*userMedia.Date)+"#"+userMedia.Key.MediaIdentifier, &userMedia)
 		if userMedia.Key.Username != username {
 			err := errors.New("Username mismatch")
 			log.Info().Err(err).Send()
@@ -83,7 +83,7 @@ func HandleRequest(ctx context.Context, history BackfillArgs) (*dynamo_types.Bat
 		return nil, err
 	}
 
-	return &dynamo_types.BatchwriteArgs{
+	return &utils.BatchwriteArgs{
 		WriteRequests: writeRequests,
 		TableName:     "media",
 		MaxBatchSize:  25,
