+++ functions/media_info/put/main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 
-	dynamo_types "github.com/KamWithK/exSTATic-backend"
+	"github.com/KamWithK/exSTATic-backend/utils"
 )
 
 var sess *session.Session
@@ -21,15 +21,15 @@ func init() {
 	svc = dynamodb.New(sess)
 }
 
-func HandleRequest(ctx context.Context, userMediaEntry dynamo_types.UserMediaEntry) error {
+func HandleRequest(ctx context.Context, userMediaEntry utils.UserMediaEntry) error {
 	userMediaEntry.LastUpdate = time.Now().Unix()
 
-	tableKey, keyErr := dynamo_types.GetCompositeKey(userMediaEntry.Key.MediaType+"#"+userMediaEntry.Key.Username, userMediaEntry.Key.MediaIdentifier)
+	tableKey, keyErr := utils.GetCompositeKey(userMediaEntry.Key.MediaType+"#"+userMediaEntry.Key.Username, userMediaEntry.Key.MediaIdentifier)
 	if keyErr != nil {
 		return keyErr
 	}
 
-	_, updateErr := dynamo_types.UpdateItem(svc, "media", tableKey, userMediaEntry)
+	_, updateErr := utils.UpdateItem(svc, "media", tableKey, userMediaEntry)
 	if updateErr != nil {
 		return updateErr
 	}
