+++ functions/settings/get/main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 
+	"github.com/KamWithK/exSTATic-backend/utils"
 	"github.com/rs/zerolog/log"
 
 	"github.com/aws/aws-lambda-go/lambda"
@@ -11,8 +12,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
-
-	dynamo_types "github.com/KamWithK/exSTATic-backend"
 )
 
 var sess *session.Session
@@ -25,7 +24,7 @@ func init() {
 	svc = dynamodb.New(sess)
 }
 
-func HandleRequest(ctx context.Context, key dynamo_types.UserSettingsKey) (*dynamo_types.UserSettings, error) {
+func HandleRequest(ctx context.Context, key utils.UserSettingsKey) (*utils.UserSettings, error) {
 	tableKey, keyErr := dynamodbattribute.MarshalMap(key)
 	if keyErr != nil {
 		log.Error().Err(keyErr).Str("table", "settings").Interface("key", key).Msg("Could not unmarshal dynamodb key")
@@ -46,7 +45,7 @@ func HandleRequest(ctx context.Context, key dynamo_types.UserSettingsKey) (*dyna
 		return nil, errors.New("Item not found in table")
 	}
 
-	optionArgs := dynamo_types.UserSettings{}
+	optionArgs := utils.UserSettings{}
 	if unmarshalErr := dynamodbattribute.UnmarshalMap(result.Item, &optionArgs); unmarshalErr != nil {
 		log.Error().Err(unmarshalErr).Str("table", "settings").Interface("key", key).Interface("item", result.Item).Msg("Could not unmarshal dynamodb item")
 		return nil, unmarshalErr
