+++ functions/settings/put/main.go
@@ -3,14 +3,13 @@ package main
 import (
 	"context"
 
+	"github.com/KamWithK/exSTATic-backend/utils"
 	"github.com/rs/zerolog/log"
 
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbattribute"
-
-	dynamo_types "github.com/KamWithK/exSTATic-backend"
 )
 
 var sess *session.Session
@@ -23,14 +22,14 @@ func init() {
 	svc = dynamodb.New(sess)
 }
 
-func HandleRequest(ctx context.Context, options dynamo_types.UserSettings) error {
+func HandleRequest(ctx context.Context, options utils.UserSettings) error {
 	tableKey, keyErr := dynamodbattribute.MarshalMap(options.Key)
 	if keyErr != nil {
 		log.Error().Err(keyErr).Str("table", "settings").Interface("key", options.Key).Msg("Could not marshal dynamodb key")
 		return keyErr
 	}
 
-	_, updateErr := dynamo_types.UpdateItem(svc, "settings", tableKey, options)
+	_, updateErr := utils.UpdateItem(svc, "settings", tableKey, options)
 	if updateErr != nil {
 		return updateErr
 	}
